Adds mechanism for contexts (#6)
Adds a mechanism for human-readable contexts. This is done by composing an instance of the DiffList class (via context_cls argument) with a type from the new contexts module. The correct type is selected from the filename extension via a lookup table in the contexts module. If no suitable extension mapping is found, or the extensions vary between the provided filenames, DefaultContext is used which just provides the filename and line number of each diff hunk.

This change also adds an ecFlow context in the EcflowContext type, which is automatically selected for the .def file extension. This translates line numbers into the relevant ecFlow node path at that line.

This completes #2 .